restore magic resource name ordering
this restores the order that the magic wiring of LWRP / custom resources
to the cookbook_name + resource filename comes after the parsing of the
file that it is defined in.

i had moved this because i was running into issues with recursive calls
between provides and the resource_name method, which i later solved.

this restores the prior behavior, but calls provides instead of
resource_name, but relies on the implicit call to resource_name from
the first provides to be functionally identical to the prior behavior
anyway.

no idea if this actually fixes any bugs, but it was not really
intentional to change it.  stuff like chef_version_for_provides
needs to be called before calling provides, so it seems like the
behavior this fixes should have been buggy with respect to that
but it is obviously...  complicated.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>